# apiv1/data/openapi.yaml
...
openapi: 3.0.3
info:
  title: Controlled Vocabularies API - Vocabulary Catalog and
    Data Endpoints
  description: |-
    An API Catalog endpoint providing access to controlled vocabularies
    in RFC 9727 linkset format.

    This v1 API is generated from the semantic mappings of the RDF
    vocabularies (see README.csv.md) and decouples the published JSON
    from legacy CSV files, while keeping the stable catalog URL
    https://schema.gov.it/api/vocabularies/v1.

    Use this API to:

    - discover REST APIs that provide controlled vocabularies published by
      various Italian public administrations;
    - search controlled vocabularies by metadata and concept identifiers.

    Schema formats are based on the OAS3.0 registry
    https://spec.openapis.org/registry/format/

    This document defines the data endpoints
    for a Controlled Vocabulary API.

    This content will be generated merging this
    "template" with the specific schema definition
    published by the provider in the repository.
  version: 1.0.0
  contact:
    name: Repository and API maintainer
    url: https://github.com/par-tec/dati-semantic-csv-apis
    email: roberto.polli@par-tec.it
  x-summary: >-
    This will be taken from skos:prefLabel of the ConceptScheme.
  termsOfService: https://schema.gov.it/terms
  x-api-id: a68b0d9e-0445-4f98-bd6e-269a97268d8b
# This is a public API.
# checkov:skip=CKV_OPENAPI_4
# checkov:skip=CKV_OPENAPI_5
...
servers:
- url: http://localhost:8080
...
  x-sandbox: true
...
paths:
...
  /vocabularies/{agencyId}/{keyConcept}/openapi.yaml:
    get:
      parameters:
      - $ref: '#/components/parameters/agencyIdPath'
      - $ref: '#/components/parameters/keyConceptPath'
      operationId: data.handlers.show_vocabulary_spec
      description: |-
        Retrieve the OpenAPI specification
        for a given vocabulary API identified by `agencyId` and `keyConcept`.
        The URL must be one of those
        retrieved from the `/vocabularies` endpoint.

        This OpenAPI resource includes
        the x-jsonld-context and x-jsonld-type extensions
        that enables the client to understand
        how to interpret the vocabulary data.
      security:
      - {}
      tags:
      - discover-vocabularies
      responses:
        '200':
          description: |-
            A successful response returns the OpenAPI specification
            for the vocabulary identified by `agencyId` and `keyConcept`.
          content:
            application/openapi+yaml:
              schema:
                type: object
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        'default':
          $ref: '#/components/responses/GenericError'
  #
  # Data endpoints.
  #
  /vocabularies/{agencyId}/{keyConcept}/dump:
    get:
      operationId: data.handlers.dump_vocabulary_dataset
      description: |-
        Dump the whole dataset
        for the vocabulary
        identified by `agencyId` and `keyConcept`.
        The URL must be one of those
        retrieved from the `/vocabularies` endpoint.

        Alternatively, this endpoint
        may return a 300 Redirect response
        to static files containing the whole dataset
        according to the vocabulary serialization format.
      parameters:
      - $ref: '#/components/parameters/agencyIdPath'
      - $ref: '#/components/parameters/keyConceptPath'
      security:
      - {}
      - high-volume: []
      tags:
      - dump-dataset
      responses:
        '200':
          description: |-
            A successful response returns the whole dataset
            from the vocabulary identified by `agencyId` and `keyConcept`.

            The response should be a compressed stream of
            all the dataset.

            Dump does not support filtering: it retrieves
            the whole dataset as is.
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/DumpResponse"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        'default':
          $ref: '#/components/responses/GenericError'
  /vocabularies/{agencyId}/{keyConcept}:
    get:
      summary: List all vocabulary items.
      description: |-
        Returns a list of all vocabulary items in the controlled vocabulary.

        Pagination is implemented via
        - offset
        - cursor, using the `id` of the last item in the previous page.
      operationId: data.handlers.show_items
      security:
      - {}
      - high-volume: []
      tags:
      - retrieve-data
      parameters:
      - $ref: '#/components/parameters/agencyIdPath'
      - $ref: '#/components/parameters/keyConceptPath'
      # Pagination.
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/offset'
      - $ref: '#/components/parameters/cursor'
      # Language selection.
      # - TODO
      # Filtering
      - $ref: '#/components/parameters/label'
      responses:
        '200':
          description: |-
            A successful response returns the data
            from the vocabulary identified by `agencyId` and `keyConcept`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        'default':
          $ref: '#/components/responses/GenericError'
  /vocabularies/{agencyId}/{keyConcept}/{id}:
    get:
      summary: Retrieve a single vocabulary item by its ID.
      description: |-
        Returns a single vocabulary item
        identified by its unique ID.

        The uniqueness of the ID inside the vocabulary
        should be guaranteed by the provider.

        Example:

        In a skos:ConceptScheme, each skos:Concept should have
        a unique skos:notation property value.

        If the skos:ConceptScheme has multiple levels,
        as long as the skos:notation is unique across all levels,
        the vocabulary item can be retrieved by its ID.

        There is no guarantee that the API URL respects
        the same hierarchy of the vocabulary levels.

        Example:

        ```turtle
        @base <https://example.com/vocab/> .

        <l1/A>
          a skos:Concept ;
          skos:notation "A" ;
          skos:prefLabel "Level 1 - A"@it .
        <l1/A/B>
          a skos:Concept ;
          skos:notation "A.B" ;
          skos:prefLabel "Level 2 - B"@it ;
          skos:broader <l1/A> .
        ```

        results in two API resources:

        ```yaml
        GET /vocab/A
        Host: example.dom

        id: A
        label: Level 1 - A
        uri: https://example.com/vocab/l1/A

        GET /vocab/A.B
        Host: example.dom

        HTTP/1.1 200 OK
        id: A.B
        label: Level 2 - B
        uri: https://example.com/vocab/l1/A/B
        parent:
        - url: https://example.com/vocab/l1/A
        ```

      operationId: data.handlers.get_item
      security:
      - {}
      tags:
      - retrieve-data
      parameters:
      - $ref: '#/components/parameters/agencyIdPath'
      - $ref: '#/components/parameters/keyConceptPath'
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-Za-z0-9._-]{1,128}$'
          maxLength: 128
        description: The unique identifier of the vocabulary item.
      responses:
        '200':
          description: |-
            A successful response returns the vocabulary item
            identified by the given ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        'default':
          $ref: '#/components/responses/GenericError'
components:
  responses:
...
    TooManyRequests:
      $ref: 'components.oas3.yaml#/components/responses/TooManyRequests'
    ServiceUnavailable:
      $ref: 'components.oas3.yaml#/components/responses/ServiceUnavailable'
    GenericError:
      $ref: 'components.oas3.yaml#/components/responses/GenericError'
  securitySchemes:
    high-volume:
      description: |-
        High volume users use this security scheme.
      type: http
      scheme: Basic
  parameters:
    limit:
      $ref: 'components.oas3.yaml#/components/parameters/limit'
    offset:
      $ref: 'components.oas3.yaml#/components/parameters/offset'
    cursor:
      name: cursor
      in: query
      explode: false
      description: |-
        Cursor for pagination.
        It is the `id` of the last item
        in the previous page.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
        maxLength: 128
      example: "11160204"
    agencyIdPath:
      $ref: 'components.oas3.yaml#/components/parameters/agencyIdPath'
    keyConceptPath:
      $ref: 'components.oas3.yaml#/components/parameters/keyConceptPath'
    label:
      name: label
      in: query
      explode: false
      description: |-
        Filter items by label.
      schema:
        $ref: '#/components/schemas/TextFilter'
      example: Idrogeno solforato
...
  schemas:
...
    TextFilter:
      description: |-
        A text filter that constrains
        the length and characters of the input.

        This is a substring match on the corresponding field of
        the vocabulary metadata.
      type: string
      maxLength: 128
      pattern: '^[A-Za-z0-9 ._:/-]{1,128}$'
    DumpResponse:
      type: string
      format: binary
...
    Limit:
      $ref: 'components.oas3.yaml#/components/schemas/Limit'
    Offset:
      $ref: 'components.oas3.yaml#/components/schemas/Offset'
...
    PaginatedResponse:
      type: object
      additionalProperties: false
      x-jsonld-type: ConceptScheme
      x-jsonld-context:
        "@vocab": "http://www.w3.org/2004/02/skos/core#"
        l0: https://w3id.org/italia/onto/l0/
        limit: null
        offset: null
        totalResults: null
        "items":
          "@container": "@set"
          "@id": l0:hasMember
      properties:
        totalResults:
          $ref: "#/components/schemas/Offset"
        limit:
          $ref: "#/components/schemas/Limit"
        offset:
          $ref: "#/components/schemas/Offset"
        items:
          type: array
          minItems: 0
          maxItems: 200
          items:
            $ref: "#/components/schemas/Item"
    Item:
      type: object
      description: |-
        A base item schema that can be extended
        with additional properties.

        This is useful for vocabularies that have a common structure
        but different properties.

        The actual schema is provided by the real OAS published
        on /{agencyId}/{keyConcept}/openapi.yaml, which extends this base schema.
      example:
        id: "11160204"
        label: Idrogeno solforato
        level: 3
        parent:
        - url: https://w3id.org/italia/work-accident/controlled-vocabulary/adm_serv/agente_causale/gruppo/1.1
        url: https://w3id.org/italia/work-accident/controlled-vocabulary/adm_serv/agente_causale/agente/11160204
        vocab:
        - url: https://w3id.org/italia/work-accident/controlled-vocabulary/adm_serv/agente_causale
      properties:
        id:
          minLength: 1
          pattern: ^[A-Za-z0-9._-]+$
          type: string
        label:
          maxLength: 500
          minLength: 1
          type: string
        parent:
          items:
            properties:
              url:
                format: uri-reference
                type: string
            required:
            - url
            type: object
          type: array
          maxItems: 100
          minItems: 0
        url:
          format: uri-reference
          type: string
        vocab:
          items:
            properties:
              url:
                format: uri-reference
                type: string
            required:
            - url
            type: object
          type: array
          maxItems: 100
          minItems: 0
      required:
      - id
      - label
      - url
      # - vocab  # vocab is not required because some items may not have the skos:inScheme property.
      x-jsonld-context:
        "@vocab": https://w3id.org/italia/work-accident/controlled-vocabulary/adm_serv/
        INAILService: https://w3id.org/italia/work-accident/onto/adm_serv/
        clvapit: https://w3id.org/italia/onto/CLV/
        dct: http://purl.org/dc/terms/
        skos: http://www.w3.org/2004/02/skos/core#
        xkos: http://rdf-vocabulary.ddialliance.org/xkos#
        xsd: http://www.w3.org/2001/XMLSchema#
        id:
          "@id": skos:notation
        label:
          "@id": skos:prefLabel
          "@language": it
        parent:
          "@container": "@set"
          "@context":
            url: "@id"
          "@id": skos:broader
        url: "@id"
        vocab:
          "@container": "@set"
          "@context":
            url: "@id"
          "@id": skos:inScheme
...